requirements.txt
Django>=3.2.5
djangorestframework>=3.12.4
psycopg2-binary>=2.9.1
pydantic[dotenv]>=1.8.2
django-money>=2.0.1
drf-yasg>=1.20.0
celery[redis]>=5.1.2